refactor(arch): exclude server-only repositories from client-reachable barrels (P0)
- Add import 'server-only' to shared/db/schema.ts and shared/db/types.ts so
  any client-bundle leak is caught at build time instead of silently bundled.
- Remove DrizzleSessionRepository, bcryptPasswordHasher, bcryptPasswordVerifier
  from session/index.ts barrel; update 6 server-action consumers to deep-import
  from @/entities/session/api and @/entities/session/lib/bcrypt.
- Remove DrizzleUserApiKeyRepository, LlmApiKeyDecryptionFailedError from
  api-key/index.ts barrel; update 5 server consumers (actions + byokGate) to
  deep-import from @/entities/api-key/api.
- Remove DrizzleContactRepository value re-export from inquiry/index.ts barrel
  (ContactInput/ContactRepository type exports retained); update action + tests.
- Add import 'server-only' to news-article/api.ts; remove DrizzleNewsRepository
  and getNewsList from news-article/index.ts barrel (NewsRow type export retained);
  update 5 action files, 2 page files, and 1 analysis action.
- Step 7 (market-news, notice) already correct — deferred by design.
- Update all affected test vi.mock paths to match new deep-import locations.
- Add byokGate.ts to ESLint no-restricted-imports ignores (pre-existing
  documented shared→entities exception, now needs /api deep path).
- Update entities/CLAUDE.md barrel exclusion table with all new entries.
