+++ src/app/[symbol]/overall/__tests__/page.body.test.tsx
@@ -33,8 +33,10 @@ vi.mock('@/shared/cache/staticSymbolCache', () => ({
 }));
 
 vi.mock('@/entities/news-article', () => ({
-    getNewsList: vi.fn().mockResolvedValue([]),
     NEWS_LIST_CACHE_KEY: 'news-list',
+}));
+vi.mock('@/entities/news-article/api', () => ({
+    getNewsList: vi.fn().mockResolvedValue([]),
 }));
 
 vi.mock('@/widgets/overall/OverallContent', () => ({
