+++ src/entities/api-key/__tests__/actions/deleteApiKeyAction.test.ts
@@ -17,7 +17,7 @@ vi.mock('next/navigation', () => ({
         throw new Error(`NEXT_REDIRECT:${path}`);
     }),
 }));
-vi.mock('@/entities/api-key', () => ({
+vi.mock('@/entities/api-key/api', () => ({
     DrizzleUserApiKeyRepository: vi.fn().mockImplementation(function () {
         return {
             deleteByUserAndProvider: mockDeleteByUserAndProvider,
