+++ src/entities/api-key/index.ts
@@ -1,7 +1,6 @@
-export {
-    DrizzleUserApiKeyRepository,
-    LlmApiKeyDecryptionFailedError,
-} from './api';
+// DrizzleUserApiKeyRepository と LlmApiKeyDecryptionFailedError は barrel から除外 —
+// api.ts が drizzle/encryption を import するため client bundle に入ると build が壊れる。
+// server consumer は @/entities/api-key/api から直接 deep import する。
 
 export { isLlmProvider, normalizeLlmApiKey } from './lib/apiKey';
 export {
