+++ src/entities/inquiry/index.ts
@@ -1,8 +1,7 @@
-export {
-    DrizzleContactRepository,
-    type ContactInput,
-    type ContactRepository,
-} from './api';
+// DrizzleContactRepository は barrel から除外 —
+// api.ts が drizzle/schema を import するため client bundle に入ると build が壊れる。
+// server consumer は @/entities/inquiry/api から直接 deep import する。
+export type { ContactInput, ContactRepository } from './api';
 
 export {
     CONTACT_CONTENT_MAX_LENGTH,
