+++ src/entities/news-article/__tests__/api.test.ts
@@ -9,7 +9,8 @@ vi.mock('@/shared/lib/sleep', () => ({
 
 import type { NewsCardAnalysis, NewsItem } from '@y0ngha/siglens-core';
 import type { SiglensDatabase } from '@/shared/db/types';
-import { DrizzleNewsRepository, type NewsRow } from '@/entities/news-article';
+import { DrizzleNewsRepository } from '@/entities/news-article/api';
+import type { NewsRow } from '@/entities/news-article';
 
 const baseItem: NewsItem = {
     id: 'abc123',
