+++ src/features/auth-logout/__tests__/actions/logoutAction.test.ts
@@ -10,9 +10,6 @@ vi.mock('@/shared/db/client', () => ({
     resetDatabaseClientForTests: vi.fn(),
 }));
 vi.mock('@/entities/session', () => ({
-    DrizzleSessionRepository: vi.fn().mockImplementation(function () {
-        return {};
-    }),
     AUTH_SESSION_COOKIE_NAME: 'siglens_session',
     applyAuthCookie: vi.fn((c: unknown) => c),
     isSecureCookieEnv: vi.fn(() => false),
@@ -20,6 +17,11 @@ vi.mock('@/entities/session', () => ({
         name: 'auth_hint',
         value: '',
     })),
+}));
+vi.mock('@/entities/session/api', () => ({
+    DrizzleSessionRepository: vi.fn().mockImplementation(function () {
+        return {};
+    }),
 }));
 // getAuthDatabaseClient는 barrel이 아닌 @/entities/session/lib/db에서 직접 import되므로
 // (server-only 체인을 client 번들에서 분리) 해당 경로를 별도로 mock한다.
