+++ src/features/auth-oauth-consent/__tests__/actions/finalizeOAuthSignupActionBranches.test.ts
@@ -33,6 +33,8 @@ vi.mock('@/entities/session', () => ({
     createAuthSession: vi.fn(),
     DEFAULT_SESSION_TTL_SECONDS: 7776000,
     isSecureCookieEnv: vi.fn(() => false),
+}));
+vi.mock('@/entities/session/api', () => ({
     DrizzleSessionRepository: vi.fn().mockImplementation(function () {
         return {};
     }),
