+++ src/shared/db/schema.ts
@@ -1,3 +1,5 @@
+import 'server-only';
+
 import { sql } from 'drizzle-orm';
 import {
     boolean,
